+++ dtbase/models/arima/run_locally.py
@@ -6,10 +6,10 @@ import pandas as pd
 import logging, coloredlogs
 
 
-from dtbase.data_utils.dataprocessor.get_data import get_training_data
-from dtbase.data_utils.dataprocessor.clean_data import clean_data
-from dtbase.data_utils.dataprocessor.prepare_data import prepare_data
-from dtbase.data_utils.config import config
+from dtbase.models.utils.dataprocessor.get_data import get_training_data
+from dtbase.models.utils.dataprocessor.clean_data import clean_data
+from dtbase.models.utils.dataprocessor.prepare_data import prepare_data
+from dtbase.models.utils.config import config
 from dtbase.models.arima.arima.arima_pipeline import arima_pipeline
 
 OUTPUT_DIR = os.path.join(os.getcwd(), "results")
