+++ dtbase/models/arima/run_pipeline.py
@@ -24,10 +24,10 @@ from dtbase.core.sensors import (
 from dtbase.models.utils.db_utils import (
     get_sqlalchemy_session,
 )
-from dtbase.data_utils.dataprocessor.get_data import get_training_data
-from dtbase.data_utils.dataprocessor.clean_data import clean_data
-from dtbase.data_utils.dataprocessor.prepare_data import prepare_data
-from dtbase.data_utils.config import config
+from dtbase.models.utils.dataprocessor.get_data import get_training_data
+from dtbase.models.utils.dataprocessor.clean_data import clean_data
+from dtbase.models.utils.dataprocessor.prepare_data import prepare_data
+from dtbase.models.utils.config import config
 from dtbase.models.arima.arima.arima_pipeline import arima_pipeline
